Make minor adjustments to map method
Following hrzn's advice, various minor adjustments were made:
* Correct various spacing/wording inconsistencies
* Move contents of test_map.py into test_timeseries.py
* Added sanity test case to tests for map
* Changed type hints to np.number
* Removed an unncessary call to self.pd_dataframe()
